tests/wpt/web-platform-tests/css/css-backgrounds/parsing/border-image-width-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-image-width with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-image-width">
<meta name="assert" content="border-image-width supports only the grammar '[ <length-percentage> | <number> | auto ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("border-image-width", "none");

test_invalid_value("border-image-width", "-1px");
test_invalid_value("border-image-width", "-2%");
test_invalid_value("border-image-width", "-3");

test_invalid_value("border-image-width", "1 2 3 4 5");
</script>
</body>
</html>
